requirements.txt
fastapi
uvicorn
langchain
langchain-community
chromadb
beautifulsoup4
requests
sentence-transformers
streamlit
python-dotenv